add AbstractCanalInstance and remove duplicate code
